chore(*): some typings
