+++ packages/wix-storybook-utils/src/typings/config.d.ts
@@ -1,0 +1,5 @@
+type Config = {
+  importFormat: string;
+  moduleName: string;
+  repoBaseURL: string;
+};
