+++ packages/wix-storybook-utils/src/typings/metadata.d.ts
@@ -1,0 +1,10 @@
+// TODO: not yet full definition
+type Metadata = {
+  displayName: string;
+  props: {
+    [s: string]: Prop;
+  };
+  readme: string;
+  readmeTestkit: string;
+  readmeAccessibility: string;
+};
