+++ packages/wix-storybook-utils/src/typings/prop.d.ts
@@ -1,0 +1,9 @@
+// TODO: not yet full definition
+type Prop = {
+  type: {
+    name: string;
+    value: [];
+  };
+  required: boolean;
+  description: string;
+};
