+++ packages/wix-storybook-utils/tsconfig.json
@@ -14,6 +14,7 @@
   },
   "include": [
     "./src/**/*.ts",
-    "./src/**/*.tsx"
+    "./src/**/*.tsx",
+    "./src/typings/*.d.ts"
   ]
 }
